Remove default minFreq and move service monitor into error monitor file
