+++ error-issue/src/error_monitor.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {ErrorReport, Stackdriver} from 'error-issue-bot';
+import {ErrorReport, ServiceGroup, Stackdriver} from 'error-issue-bot';
 import {StackdriverApi} from './stackdriver_api';
 
 import fetch from 'node-fetch';
@@ -24,12 +24,41 @@ export const ERROR_ISSUE_ENDPOINT =
   process.env.ERROR_ISSUE_ENDPOINT ||
   'https://us-central1-amp-error-issue-bot.cloudfunctions.net/error-issue';
 
+export enum ServiceName {
+  PRODUCTION = 'CDN Production',
+  DEVELOPMENT = 'CDN Development',
+  EXPERIMENTS = 'CDN Experiments',
+  NIGHTLY = 'CDN Nightly',
+}
+
+/**
+ * Set of service groups to monitor and details for scaling frequency.
+ * Note: These values do not need to be exact; the order-of-magnitude is what's
+ * important here.
+ */
+const SERVICE_GROUPS: Record<ServiceName, ServiceGroup> = {
+  'CDN Production': {diversionPercent: 0.98, throttleRate: 0.1},
+  'CDN Development': {diversionPercent: 0.01, throttleRate: 1},
+  'CDN Experiments': {diversionPercent: 0.01, throttleRate: 1},
+  'CDN Nightly': {diversionPercent: 0.0005, throttleRate: 1},
+};
+
 export class ErrorMonitor {
   constructor(
     protected client: StackdriverApi,
-    readonly minFrequency: number = 5000,
+    readonly minFrequency: number,
     protected pageLimit: number = 25
   ) {}
+
+  /** Creates a service monitor using the same settings as this monitor. */
+  service(serviceName: ServiceName): ServiceErrorMonitor {
+    return new ServiceErrorMonitor(
+      this.client,
+      serviceName,
+      this.minFrequency,
+      this.pageLimit
+    );
+  }
 
   /** Tests if an error group already has an associated issue. */
   private hasTrackingIssue(group: Stackdriver.ErrorGroup): boolean {
@@ -121,5 +150,36 @@ export class ErrorMonitor {
     }
 
     return urls;
+  }
+}
+
+/**
+ * Returns the scaling factor to normalize frequency for a service group against
+ * what it would be in production traffic.
+ */
+function scaleFactor(serviceName: ServiceName): number {
+  const {
+    diversionPercent: prodPercent,
+    throttleRate: prodThrottle,
+  } = SERVICE_GROUPS[ServiceName.PRODUCTION];
+  const {diversionPercent, throttleRate} = SERVICE_GROUPS[serviceName];
+  return (prodPercent * prodThrottle) / (diversionPercent * throttleRate);
+}
+
+export class ServiceErrorMonitor extends ErrorMonitor {
+  // Note that minFrequency is relative to production traffic, and is scaled for
+  // each diversion when thresholding.
+  constructor(
+    client: StackdriverApi,
+    private serviceName: ServiceName,
+    minFrequency: number,
+    pageLimit = 25
+  ) {
+    super(client, minFrequency / scaleFactor(serviceName), pageLimit);
+  }
+
+  /** Finds top occurring errors in the service group. */
+  protected async newErrors(): Promise<Array<Stackdriver.ErrorGroupStats>> {
+    return this.client.listServiceGroups(this.serviceName, this.pageLimit);
   }
 }
